% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith19.R
\docType{data}
\name{smith19.use}
\alias{smith19.use}
\title{Smith19.use data.}
\format{
A data frame with 3444 observations and 39 variables
}
\usage{
smith19.use
}
\description{
A dataset is used in Parenting Risk and Resilience, Social-emotional Readiness,
 and Reading Achievement in Kindergarten Children from Low-income Families Model.
 Author(s): Sondra Smith-Adcock, Walter Leite, Yasemine Kaya & Ellen Amatea.
 **Source:** Journal of Child and Family Studies, vol 28(Oct., 2019), pp. 2826-2841.
 Published by: Springer Nature in Journal of Child and Family studies
 DOI:<https://doi.org/10.1007/s10826-019-01462-0>
}
\details{
The dataset was taken from the public-use data of
 the Early Childhood Longitudinal Study – Kindergarten Class of 1998-99 of
 the National Center for Educational Statistics (https://nces.ed.gov/ecls/kindergarten.asp/).
 This dataset should not be combined with other data for the purpose of identifying participants.
}
\keyword{datasets}
