% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens.tables.R
\name{sens.tables}
\alias{sens.tables}
\title{Summary of sensitivity analysis results}
\usage{
sens.tables(expr = NULL, sig.level = 0.05, choice = NULL)
}
\arguments{
\item{expr}{Returned object of \code{\link{sa.aco}} function.}

\item{sig.level}{Significance level, default value is 0.05.}

\item{choice}{Set up the length of summary; default is all.}
}
\value{
List of summary tables
}
\description{
This function can summarize the sensitivity analysis results from
   \code{\link{sa.aco}} function.
}
\examples{
# see examples in the \code{\link{sa.aco}} function

}
